+++ packages/docs/src/DocHome.js
@@ -1,7 +1,8 @@
-import React, {useCallback, useEffect, useState} from 'react';
-import {Route, Switch} from 'react-router-dom';
-import {initStore, RouteLoader, StoreContext} from 'react-windy-ui';
+import React, {useCallback, useContext, useEffect, useState} from 'react';
+import {Redirect, Route, Switch, useHistory} from 'react-router-dom';
+import {CssThemeProvider, initStore, Loader, RouteLoader, StoreContext} from 'react-windy-ui';
 import intl from 'react-intl-universal';
+import {DocThemeContext} from "./common/DocConstants";
 
 // locale data
 const locales = {
@@ -22,13 +23,14 @@ const barStyle = {
 };
 
 const Loading = () => {
-  return 'Loading the page...';
+  // return 'Loading the page...';
+  return  <Loader type="third" active={true} size="small" style={{marginRight: '1rem'}}/>;
 };
 
 export default function DocHome() {
   const [lang, setLang] = useState({initDone: false, currentLocale: 'zh_CN'});
   const [store] = useState(() =>
-      initStore('init'),
+    initStore('init'),
   );
 
   const switchLocale = useCallback((lc) => {
@@ -45,41 +47,49 @@ export default function DocHome() {
   }, [lang.currentLocale, switchLocale]);
 
   const changeLocale = useCallback((nextLang) => {
-    switchLocale(nextLang).
-        then(() => setLang(pre => ({...pre, currentLocale: nextLang})));
+    switchLocale(nextLang).then(() => setLang(pre => ({...pre, currentLocale: nextLang})));
   }, [switchLocale]);
+
+  const {theme, change} = useContext(CssThemeProvider.context);
+  const history = useHistory();
+  const setTheme = useCallback((themeName) => {
+    change(themeName);
+    history.push(`/docs/${themeName}`);
+  }, [change, history]);
 
   return <>
     <React.Suspense fallback={<Loading/>}>
-      <StoreContext.Provider
+      <DocThemeContext.Provider value={{theme, setTheme}}>
+        <StoreContext.Provider
           value={{
             supportLocals: Object.keys(locales),
             locale: lang.currentLocale,
             changeLocale,
             store,
           }}>
-        <Switch>
-          <RouteLoader
+          <Switch>
+            <RouteLoader
               route={Route}
-              path="/docs"
+              path={`/docs/:theme`}
               render={() => <DocCenter/>}
               progressStyle={progressStyle}
               barStyle={barStyle}>
-          </RouteLoader>
-          <RouteLoader
+            </RouteLoader>
+            <RouteLoader
               exact
               route={Route}
-              path="/"
+              path={`/:theme`}
               render={() => <Home/>}
               progressStyle={progressStyle}
               barStyle={barStyle}>
-          </RouteLoader>
+            </RouteLoader>
+            <Redirect from="/" to={`/${theme}`}/>
+            <RouteLoader route={Route} render={() => <div>404, 页面不存在~~</div>}>
 
-          <RouteLoader route={Route} render={() => <div>404, 页面不存在~~</div>}>
-
-          </RouteLoader>
-        </Switch>
-      </StoreContext.Provider>
+            </RouteLoader>
+          </Switch>
+        </StoreContext.Provider>
+      </DocThemeContext.Provider>
     </React.Suspense>
   </>;
 }