+++ packages/docs/src/common/DocConstants.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const defaultTheme = 'default';
+
+export const DocThemeContext = React.createContext(null);