+++ packages/docs/src/home/HomeFooter.js
@@ -6,62 +6,19 @@ import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 export default function HomeFooter() {
   return <div className="doc home-footer">
     <Row>
-      <Col xs={12} sm={3}>
-        <div className="doc footer-panel">
-          <div className="doc footer-title">
-            其他资源
-          </div>
-          <ul className="doc home-link">
-            <li>代码仓库</li>
-            <li>Wui样式库</li>
-            <li>图标库</li>
-            <li>在线主题编辑器</li>
-          </ul>
-        </div>
-
-      </Col>
-      <Col xs={12} sm={3}>
-        <div className="doc footer-panel">
-          <div className="doc footer-title">
-            成为贡献者
-          </div>
-          <ul className="doc home-link">
-            <li>提交更新</li>
-            <li>报告Bug</li>
-            <li>修改样式</li>
-          </ul>
-        </div>
-      </Col>
-      <Col xs={12} sm={3}>
-        <div className="doc footer-panel">
-          <div className="doc footer-title">
-            获取帮助
-          </div>
-          <ul className="doc home-link">
-            <li>常见问题</li>
-            <li>在线讨论</li>
-            <li>StackOverflow</li>
-          </ul>
-        </div>
-      </Col>
-      <Col xs={12} sm={3}>
-        <div className="doc footer-panel right">
-          <div className="doc footer-title">
-            进行捐助
-          </div>
-          <Popover position="top" autoWidth body={<div>
-            <p>Haha, you're so kind. </p>
-            <p>Appreciate to access this project.</p>
-            {<FontAwesomeIcon icon={faSmile} style={{color: '#49b847'}}/>}
-            {<FontAwesomeIcon icon={faSmile} style={{color: '#49b847'}}/>}
-            {<FontAwesomeIcon icon={faSmile} style={{color: '#49b847'}}/>}
-          </div>}>
-            <Button color="green" outline>我要捐助</Button>
-          </Popover>
-        </div>
-      </Col>
       <Col col={12} flexCol style={{fontSize: '.8rem'}}>
-        <a href="https://beian.miit.gov.cn/" target="_blank">苏ICP备18064205号</a>
+        <a href="https://beian.miit.gov.cn/" target="_blank" rel="noreferrer"
+           style={{marginRight: '1rem'}}>苏ICP备18064205号</a>
+
+        <Popover position="top" autoWidth body={<div>
+          <p>Haha, you're so kind. </p>
+          <p>Appreciate to access this project.</p>
+          {<FontAwesomeIcon icon={faSmile} style={{color: '#49b847'}}/>}
+          {<FontAwesomeIcon icon={faSmile} style={{color: '#49b847'}}/>}
+          {<FontAwesomeIcon icon={faSmile} style={{color: '#49b847'}}/>}
+        </div>}>
+          <Button color="white" outline size="small">我要捐助</Button>
+        </Popover>
       </Col>
     </Row>
   </div>;